requirements.txt
pytest==7.2.0
selene>=2.0.0rc2
allure-pytest==2.13.2